+++ src/metatrain/experimental/lorem/modules/__init__.py
@@ -1,1 +1,13 @@
+from .backbone import LoremBackbone
+from .bec import BornEffectiveChargeHead
+from .long_range import LoremLongRangeFeaturizer
+from .tensor_dense import TensorDense, TensorProduct
 
+
+__all__ = [
+    "BornEffectiveChargeHead",
+    "LoremBackbone",
+    "LoremLongRangeFeaturizer",
+    "TensorDense",
+    "TensorProduct",
+]